Attempt to break the binary GC test
I'm still not 100% convinced that I have the source script reference
handling correct. I'm pretty sure I need to be keeping the underlying
binary around for the lifetime of the ExternalArrayBuffer.

However, this should be straightforward by using the
JsSetObjectBeforeCollectCallback call to attach a callback function to
the ExternalArrayBuffer to free the underlying storage when the buffer
is garbage collected.