Allow for multiple contexts per runtime
This is a large commit so apologies for the churn. This creates a new
resource type for runtimes that is separate from the context resource.
This way users are able to create multiple contexts per runtime which
allows for multiple self contained contexts on a given runtime. The
reason for this is due to the runtime being more heavy on resource
consumption due to its default use of extra threads and so on.