Bug fixes
So I fixed three related things at once and I'm not going to bother to
separate the commit. Sue me.

1. Run enif_inspect_binary in each callback

Apparently I can't use an ErlNifBinary between calls to NIF functions as
things tend to die with weird memory errors. Not a big hassle but ended
up being a decent sized diff.

2. Explicitly check disabled state

Some calls (JsSerialize) will trigger assertion errors with a debug
library if called on a runtime that has been disabled. This just adds an
explicit check everywhere so that we return the error rather than crash
the process.

3. Replace JsRun with JsSerialize/JsRunSerialized

It turns out that Chakra does want to hold onto a copy of the script
that's passed through JsRun. This seemed to work until using a debug
build of ChakraCore. There's no enif_keep_binary or equivalent and
holding large binaries is done by creating a resource that holds an
unbound ErlNifEnv with a copy of the term. Which it turns out is exactly
what our script serialization already does. So rather than duplicate
this work I just removed the direct call to JsRun and wrote five lines
of Erlang to do the equivalent. This also has the happy benefit of
testing serialization a whole lot more and removes the number of ways to
invoke JS in C.