Enable message passing to and from JavaScript
This allows for async communication between Erlang and JavaScript by
providing the following three built-in JavaScript functions:

  * `erlang.has_message()` - returns true or false
  * `erlang.receive()` - returns the message (undefined if conversion
    fails)
  * `erlang.send(To, Message)` - returns undefined or throws an exception

Along with the following Erlang API:

  * `chakra:send(Context, Message)`

There are a few caveats to the use of this API. First, any JavaScript
that calls `erlang.receive()` will block the Erlang scheduler until the
script is interupted so if its expected to have untrusted user
JavaScript running then its imperative that script interruption is used
to prevent a DOS on dirty CPU schedulers.

Secondly, there is a single message queue per _Runtime_. This means that
if you send messages to a specific context, and do not consume them in
that context, an unrelated context on the same runtime will end up
consuming those messages.

Hopefully I'll be able to address both of these issues.