WIP: Massive rewrite in progress
This is changing away from using dirty scheduler threads to using a
thread per runtime. Dirty schedulers have a fairly low default stack
size and it would be a bummer to force everyone that uses Chakra to have
to increase that limit (which isn't always straightforward).

This new architecture should also allow for much cleaner RPC style
interaction between JavaScript and Erlang so that's also nice.